pyproject.toml
[tool.poetry]
name = "sentineliq"
version = "0.1.0"
description = "Sentineliq - A B2B Django REST API"
authors = ["SentinelIQ"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = ">=5.0"
djangorestframework = ">=3.15"
djangorestframework-simplejwt = ">=5.3"
psycopg2-binary = ">=2.9.9"
python-dotenv = ">=1.0.0"
gunicorn = ">=21.2.0"
drf-spectacular = ">=0.28.0"
django-filter = ">=25.1"
django-debug-toolbar = ">=4.3.0"
django-cors-headers = "^4.7.0"
weasyprint = "^65.1"
markdown2 = {version = ">=2.5.3,<3.0.0", markers = "python_version < \"4.0.0\""}
redis = "^6.0.0"
celery = "^5.5.2"
django-celery-beat = "^2.8.0"
kombu = "^5.5.3"
pika = "^1.3.2"
sentry-sdk = {extras = ["celery", "django"], version = "^2.27.0"}
flower = "^2.0.1"
tornado = "^6.4.2"
django-model-utils = "^5.0.0"
idna = "^3.10"
elasticsearch-dsl = "^8.18.0"

[tool.poetry.group.dev.dependencies]
black = ">=24.2.0"
isort = ">=5.13.2"
flake8 = ">=7.0.0"
pytest = ">=8.0.0"
pytest-django = ">=4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
